+++ projects/ship/src/main/java/org/openmuc/jeebus/ship/api/ConfigBuilder.java
@@ -19,10 +19,7 @@ import org.slf4j.LoggerFactory;
 
 import java.lang.invoke.MethodHandles;
 import java.net.*;
-import java.util.Arrays;
-import java.util.Objects;
-import java.util.Optional;
-import java.util.Set;
+import java.util.*;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
@@ -48,8 +45,8 @@ import static java.nio.charset.StandardCharsets.UTF_8;
  * the required parameters.
  */
 public final class ConfigBuilder {
-    public static final InetSocketAddress IP_V4_ANY;
-    public static final InetSocketAddress IP_V6_ANY;
+    public static final InetSocketAddress INET4_ANY;
+    public static final InetSocketAddress INET6_ANY;
 
     private static final Logger LOG = LoggerFactory.getLogger(
         MethodHandles.lookup().lookupClass()
@@ -57,11 +54,11 @@ public final class ConfigBuilder {
 
     static {
         try {
-            IP_V4_ANY = new InetSocketAddress(
+            INET4_ANY = new InetSocketAddress(
                 Inet4Address.getByAddress(new byte[] { 0, 0, 0, 0 }),
                 0
             );
-            IP_V6_ANY = new InetSocketAddress(
+            INET6_ANY = new InetSocketAddress(
                 InetAddress.getByName("::"),
                 0
             );
@@ -75,9 +72,12 @@ public final class ConfigBuilder {
     private String id;
 
     private boolean serverEnabled = true;
-    private Set<InetSocketAddress> serverBindAddresses = Set.of(IP_V4_ANY);
+    private Set<InetSocketAddress> serverBindAddresses = Set.of(INET4_ANY);
     private boolean autoAcceptEnabled = false;
-    private Set<String> trustedSkis;
+    private Set<String> trustedSkis = Collections.emptySet();
+
+    private long networkInterfaceScanInterval = 10;
+    private long networkInterfaceScanInitialDelay = 5;
 
     private String mDnsServiceInstance;
     private String mDnsDomain = "local.";
@@ -126,9 +126,15 @@ public final class ConfigBuilder {
      * service discovery using mDNS. This includes finding remote SHIP services as
      * well as registering the service of this device.
      * <p>
-     * By default, this is set to {@code 0.0.0.0:0} to consider all available
-     * network interfaces and use ephemeral free ports.
-     * {@code [::]:0} does the same.
+     * By default, this is set to the ANY adress ({@code 0.0.0.0:0}) to consider all
+     * available network interfaces and use ephemeral free ports. {@code [::]:0} does
+     * the same. If set to the ANY address, the SHIP node will also scan for new
+     * network interfaces during runtime to bind its server to and register new mDNS
+     * services. Else it will only consider addresses given here.
+     * <p>
+     * Support for multiple socket addresses with different port numbers is limited
+     * and can lead to wrong mDNS services being registered or discovered. It is
+     * recommended to choose one free port number and use that for all sockets.
      *
      * @param serverBindAddresses
      *     Socket addresses to bind the SHIP server to. IPv4 and IPv6 addresses are
@@ -164,6 +170,9 @@ public final class ConfigBuilder {
      * {@code "127.0.0.1:8080"};
      * {@code "[1080:0:0:0:8:800:200C:417A]:2001"};
      * {@code "[1080::8:800:200C:417A]:6060"}
+     * <p>
+     * Please see {@link ConfigBuilder#withServerBindAddresses(Set)} for further
+     * important information on socket address configuration.
      *
      * @param serverBindAddresses
      *     String representations of the socket address to bind the SHIP server to
@@ -202,6 +211,42 @@ public final class ConfigBuilder {
      */
     public ConfigBuilder withAutoAcceptEnabled(boolean autoAcceptEnabled) {
         this.autoAcceptEnabled = autoAcceptEnabled;
+        return this;
+    }
+
+    /**
+     * For mDNS services and the SHIP server to react to changes in the available
+     * network interfaces jEEBus.SHIP does periodic scans and refreshes its server,
+     * services and service listeners accordingly. Here you can configure the
+     * interval between these scans in seconds.
+     *
+     * @param intervalInSeconds
+     *     the interval between network interface scans
+     * @return the updated {@link ConfigBuilder}
+     * @see ConfigBuilder#withNetworkInterfaceScanInitialDelay(long)
+     */
+    public ConfigBuilder withNetworkInterfaceScanInterval(long intervalInSeconds) {
+        this.networkInterfaceScanInterval = intervalInSeconds;
+        return this;
+    }
+
+    /**
+     * For mDNS services and the SHIP server to react to changes in the available
+     * network interfaces jEEBus.SHIP does periodic scans and refreshes its server,
+     * services and service listeners accordingly. Here you can configure an initial
+     * delay before the scanning starts in seconds. Defaults to {@code 5} seconds.
+     * This will also delay mDNS scanning as a whole to give the SHIP node some time
+     * to initialize before remote SHIP services are reported.
+     *
+     * @param initialDelayInSeconds
+     *     the initial delay before network interface scans
+     * @return the updated {@link ConfigBuilder}
+     * @see ConfigBuilder#withNetworkInterfaceScanInterval(long)
+     */
+    public ConfigBuilder withNetworkInterfaceScanInitialDelay(
+        long initialDelayInSeconds
+    ) {
+        this.networkInterfaceScanInitialDelay = initialDelayInSeconds;
         return this;
     }
 
@@ -422,33 +467,20 @@ public final class ConfigBuilder {
         if (theAnyAddress.isPresent()) {
             LOG.info(
                 "SHIP server bind address is set to the ANY address. Using all"
-                    + " network interfaces supporting multicast for server binding"
+                    + " available network interfaces for server binding"
                     + " and mDNS Service Discovery."
             );
-            try {
-                this.serverBindAddresses = NetworkInterface
-                    .networkInterfaces()
-                    .filter(this::safelyCheckMulticastSupport)
-                    .flatMap(NetworkInterface::inetAddresses)
-                    .map(addresss -> new InetSocketAddress(
-                        addresss,
-                        theAnyAddress.get().getPort()
-                    ))
-                    .collect(Collectors.toSet());
-            }
-            catch (SocketException e) {
-                throw new IllegalArgumentException(
-                    "There was an error retrieving available network interfaces",
-                    e
-                );
-            }
         }
         return new ShipConfig(
             id,
             serverEnabled,
-            serverBindAddresses,
+            // if there is an ANY address, disregard other addresses
+            theAnyAddress.map(Set::of).orElse(serverBindAddresses),
+            theAnyAddress.isPresent(),
             autoAcceptEnabled,
             trustedSkis,
+            networkInterfaceScanInitialDelay,
+            networkInterfaceScanInterval,
             mDnsServiceInstance,
             mDnsDomain,
             brand,
@@ -487,9 +519,9 @@ public final class ConfigBuilder {
         Set<InetSocketAddress> serverBindAddresses
     ) {
         return serverBindAddresses.stream()
-            .filter(address -> IP_V4_ANY.getAddress()
-                .equals(address.getAddress())
-                || IP_V6_ANY.getAddress().equals(address.getAddress()))
+            .filter(socket ->
+                INET4_ANY.getAddress().equals(socket.getAddress())
+                || INET6_ANY.getAddress().equals(socket.getAddress()))
             .findAny();
     }