+++ projects/ship/src/main/java/org/openmuc/jeebus/ship/util/ShipUtilities.java
@@ -10,18 +10,35 @@
 
 package org.openmuc.jeebus.ship.util;
 
+import javax.annotation.Nonnull;
 import java.net.Inet6Address;
+import java.net.InetAddress;
 import java.net.InetSocketAddress;
+import java.util.Arrays;
+import java.util.Comparator;
+import java.util.TreeSet;
+import java.util.stream.Collector;
+import java.util.stream.Collectors;
 
 public class ShipUtilities {
-    @SuppressWarnings("HardcodedFileSeparator")
     public static String beautify(InetSocketAddress socket) {
-        String hostname = socket.getHostName();
-        String ip = socket.getAddress().getHostAddress();
-        int port = socket.getPort();
+        return getStringBuilder(socket.getAddress())
+            .append(":")
+            .append(socket.getPort())
+            .toString();
+    }
+
+    public static String beautify(InetAddress address) {
+        return getStringBuilder(address).toString();
+    }
+
+    @Nonnull
+    private static StringBuilder getStringBuilder(InetAddress address) {
+        String hostname = address.getHostName();
+        String ip = address.getHostAddress();
 
         boolean hasHostname = hostname != null && !ip.startsWith(hostname);
-        boolean isIPv6 = socket.getAddress() instanceof Inet6Address;
+        boolean isIPv6 = address instanceof Inet6Address;
 
         StringBuilder sb = new StringBuilder();
 
@@ -34,9 +51,27 @@ public class ShipUtilities {
         } else {
             sb.append(ip);
         }
+        return sb;
+    }
 
-        sb.append(":").append(port);
+    public static final Comparator<InetAddress> SCOPED_ADDRESS_ORDER
+        = (left, right) -> {
 
-        return sb.toString();
+            int byteCompare = Arrays.compare(left.getAddress(), right.getAddress());
+
+            if (byteCompare == 0
+                && left instanceof Inet6Address && right instanceof Inet6Address
+            ) {
+                return Integer.compare(
+                    ((Inet6Address) left).getScopeId(),
+                    ((Inet6Address) right).getScopeId()
+                );
+            }
+            return byteCompare;
+        };
+
+    @Nonnull
+    public static Collector<InetAddress, ?, TreeSet<InetAddress>> toScopedAddressTreeSet() {
+        return Collectors.toCollection(() -> new TreeSet<>(SCOPED_ADDRESS_ORDER));
     }
 }